requirements.txt
gradio>=4.0.0
transformers==4.46.3
peft==0.13.2
torch>=2.0.0
sentence-transformers
faiss-cpu
langchain-huggingface
numpy
